u64

+++ src/damselfly/update_interval/update_interval_sorter.rs
@@ -6,7 +6,7 @@ pub struct UpdateIntervalSorter;
 impl UpdateIntervalSorter {
     pub fn sort_by_timestamp(updates: &mut Vec<&UpdateInterval>) {
         updates.sort_unstable_by(|prev, next| {
-            let get_timestamp = |update_interval: &&UpdateInterval| -> usize {
+            let get_timestamp = |update_interval: &&UpdateInterval| -> u64 {
                 match &update_interval.val {
                     MemoryUpdateType::Allocation(allocation) => allocation.get_timestamp(),
                     MemoryUpdateType::Free(free) => free.get_timestamp(),
